use more precise types

+++ draw.h
@@ -90,10 +90,10 @@ typedef struct
 {
     xcb_pixmap_t pixmap;
     xcb_visualtype_t *visual;
-    int width;
-    int height;
+    uint16_t width;
+    uint16_t height;
     int phys_screen;
-    int depth;
+    uint8_t depth;
     cairo_t *cr;
     cairo_surface_t *surface;
     PangoLayout *layout;
